+++ test/fixtures/rails-with-frontend/config/application.rb
@@ -1,0 +1,6 @@
+require "rails"
+
+module FixtureRailsWithFrontend
+  class Application < Rails::Application
+  end
+end